1. Change id type to long.

+++ src/main/java/com/buttons/smarthome/repo/ApartmentRepo.java
@@ -3,5 +3,5 @@ package com.buttons.smarthome.repo;
 import com.buttons.smarthome.models.Apartment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ApartmentRepo extends CrudRepository<Apartment, Integer> {
+public interface ApartmentRepo extends CrudRepository<Apartment, Long> {
 }
